Update README.md
"Harry Potter and the Chamber of Secrets follows Harry's second year at Hogwarts School of Witchcraft and Wizardry, where a mysterious series of attacks on students leaves them petrified. Harry, Ron, and Hermione discover that the attacks are linked to the legend of the Chamber of Secrets, an ancient hidden chamber housing a deadly monster. With the help of friends and courage, Harry uncovers the truth, confronts Tom Riddle (a younger version of Voldemort), and defeats the basilisk to save Ginny Weasley and the school."

## Key Information

**Author: J.K Rowling**
**Published: July 2, 1998(UK), June 2, 1999(US)**
**Genre: Fantasy, Adventure, Mystery**
**Pages: 251(UK Edition), 341(US Edition)**